refactor: ChatWidget을 App.tsx로 이동 (#13)
