+++ src/components/organisms/ChatWidget/ChatWidget.module.css
@@ -11,6 +11,7 @@
   position: fixed;
   bottom: 1rem;
   right: 1rem;
+  z-index: 100;
 }
 .button:hover {
   background-color: var(--color-navy);
